+++ src/utils/index.ts
@@ -8,7 +8,7 @@ export * from './is';
 export * from './position';
 export * from './resize-observer-helper';
 export * from './scroll';
-export * from './style-helper';
+export * from './style';
 export * from './transformer';
 export * from './transition-event-helper';
 export * from './types';
